Crate binarray

BinArray

Uses standard unsigned integers as statically sized binary arrays. This will have improved performance over the BitVecs dynamically sized binary vector. This is done by keeping the array out of heap allocation, improving locality, and uses lower cost functions. Having a static array also permits other optimisations. To further enhance performance, the functions do not provide error checking due to potential branch prediction misses. It is the respon- sibility of the programmer to ensure the index is within bounds, for example.

Bit Manipulation Functions

get_bit() Returns the bit value at location index

set_bit() Sets the bit value at location index

to_bstring() converts the array into a padded binary string

Functions not included as they are already standard

count_ones()

count_zeros()

from_be_bytes()

from_le_bytes()

leading_ones()

leading_zeros()

to_be_bytes()

to_le_bytes()

Traits

BinaryArray